perl5db: only check sub names for non-references
DB::sub can be called by perl giving it either a name (if it can be
determined) or a code ref. There is special handling for AUTOLOAD subs
and threads::new. This could only be happen when given the name of the
sub, so there is no need to do these checks if given a reference.
Additionally, a reference could be an object, which could have
overloads. Those overloads could fail or otherwise complicate the normal
operation of DB::sub.

Add a ref check to the sub given to DB::sub to avoid these issues.