Add some ARGS_ASSERT calls
An ARGS_ASSERT macro is always generated for every function listed in
embed.fnc, unless possibly suppressed with the G flag for that entry.
The generated macro is empty if there is nothing to assert.

It is mandatory (enforced through a porting test) to call that macro
when non-empty.  (Hopefully the call occurs in the function the macro is
designed for, but the porting test is currently simplistic and doesn't
check for that; often compilation would fail anyway if it did get placed
in the wrong function, as the parameter names the macro expects and the
ones in the function could easily not match).

It is optional (but a good idea) to call the macro even when empty.
That way this commit would not have been necessary.  From time to time,
an empty macro becomes non-empty as we figure out more things to check
for.  When that happens, the porting test fails for the newly-non-empty
macros that aren't called.  If the function had originally called the
empty-one, its source wouldn't have to change at all.

Several commits from now will make some ARGS_ASSERT macros non-empty;
this commit adds calls to the ones that weren't already called.