regen/embed.pl: Add a condition to an 'if'
This is in preparation of combining this if clause with the one just
below to as many of the conditions in common as possible.

In this clause, it applies only to functions, as we don't create
ARGS_ASSERT macros for macros.  However when something is going into
mathoms.c (b flag), it is marked as a macro, but also has a function, so
we need to create the ASSERT in this case.