peep.c: Only remove expected empty if/else blocks structures
The expected OP tree structures for empty blocks fit for removal was:
```
  +-- stub
```
or
```
  +-- scope
    |
    +-- stub
```

GH #23505 showed that `{ BEGIN {} }` blocks will also generate the likes of:
```
  +--scope
    |
    +--stub
    |
    +--null (ex-nextstate)
```

This commit does the minimal possible to fix this bug - the peephole optimizer
now checks for the expected cases and ignores anything else.

Future commits may build on this to remove this additional type of structure.
Additional tests have been added now and give a useful structure for future
commit tests to fit into.