regen/embed.pl: Combine two arrays
These two arrays are combined anyway into a third array later in the
code; this just goes with the third array from the start.

The order of some assertions in the generated proto.h are changed.  But
this makes sure that we continue to assert that an argument is present
before dereferencing it.