Remove customized ARGS_ASSERT macros
These existed because they exceeded the rudimentary capabilities of the
generated ARGS_ASSERT macros.  But the previous commit generalized that
capability, so they can change to use that.

This moves the three extant customized macros to use the new facility,
removing the need for overriding things

The customized PERL_ARGS_ASSERT_ISA_LOOKUP was out-of-date, missing an
assertion that got added, without the macro getting updated.  This new
scheme will prevent that from happening in the future.