Move assert for av_new_alloc() to embed.fnc
You can't allocate a zero length array using this function; this uses
the new assert() functionality in embed.fnc introduced in the previous
few commits to cause the assertion to be placed in
PERL_ARGS_ASSERT_AV_NEW_ALLOC rather than be in the function.

There is no particular need to do this; this is mostly a demonstration
of how it can be done.