Revert "todo test for 14615 (<<>> deparses same as <>)"
This reverts commit aa1c6e98da541a26043a44ce0b075903f910f0ad.
This reverts commit 3c8eed32137140950214d1babd39e0f767ebc401.

The <<>> operator was introduced in v5.22. During development (the v5.21
branch) it was discovered that B::Deparse didn't handle it correctly.
That's GH #14615.

B::Deparse was fixed in commit 65ef2c3e7d945498f4ef8ab758cc14a8af2f7533,
but the bug ticket was left open because there was no regression test
for B::Deparse yet.

The missing regression test was added the same day in commit
2fcbb74a67c62e89b625fc824fcddb7de6519a28, but without reference to the
bug ticket, so no one noticed.

When v5.22 was released, it had a new <<>> operator, full B::Deparse
support, and a regression test for it. The only thing missing was to
close #14615.

This new test in t/run/todo.t is not just redundant given the history
above, it also breaks 'make minitest':

    # Failed test 2 - No assertion failure at run/todo.t line 50
    #      got "512"
    # expected "0"
    # Failed test 3 - while(<<>>) deparses as while (defined($_ = <<>>)) at run/todo.t line 52
    #      got 'Can\'t locate B.pm in @INC (@INC entries checked: /home/runner/work/perl5/perl5/cpan/AutoLoader/lib /home/runner/work/perl5/perl5/dist/Carp/lib /home/runner/work/perl5/perl5/dist/PathTools /home/runner/work/perl5/perl5/dist/PathTools/lib /home/runner/work/perl5/perl5/cpan/ExtUtils-Install/lib /home/runner/work/perl5/perl5/cpan/ExtUtils-MakeMaker/lib /home/runner/work/perl5/perl5/cpan/ExtUtils-Manifest/lib /home/runner/work/perl5/perl5/cpan/File-Path/lib /home/runner/work/perl5/perl5/ext/re /home/runner/work/perl5/perl5/dist/Term-ReadLine/lib /home/runner/work/perl5/perl5/dist/Exporter/lib /home/runner/work/perl5/perl5/ext/File-Find/lib /home/runner/work/perl5/perl5/cpan/Text-Tabs/lib /home/runner/work/perl5/perl5/dist/constant/lib /home/runner/work/perl5/perl5/cpan/version/lib /home/runner/work/perl5/perl5/cpan/Getopt-Long/lib /home/runner/work/perl5/perl5/cpan/Text-ParseWords/lib /home/runner/work/perl5/perl5/cpan/ExtUtils-PL2Bat/lib /home/runner/work/perl5/perl5/lib . .) at /home/runner/work/perl5/perl5/lib/B/Deparse.pm line 57.\nBEGIN failed--compilation aborted at /home/runner/work/perl5/perl5/lib/B/Deparse.pm line 57.\nCompilation failed in require at - line 1.\nBEGIN failed--compilation aborted at - line 1.'
    # expected /(?^:defined\(\$_ = <<>>\))/

Easiest solution: Just revert it.