dump.c: Fix broken c++ compilations in
Commit bce5eba453b71437b5d508b1e717646587dda0db broke c++ compilations
because a goto now crosses declarations with initialization.

This commit merely separates the declarations from their initialization,
and moves the declarations to before the goto.

The alternative could just as easily have been to move the goto to after
the initialization, but the initialization would then be (a bit of)
extra work, and more importantly, that would break the code flow of
immediately using the initialized values.