op.c: Don't hand-roll is_dup_mode() attributes
This commit adds an entry in embed.fnc for S_is_dup_mode, removing the
assert and __attribute lines in op.c.

The proximal cause for this commit is that I tried compiling with
Perl_assert() enabled, which resulted in a lot of compiler warnings
because of the __attribute__non_null__ line

But I also think it is better to not hand-roll things unless absolutely
necessary.  Changes someone makes to the general scheme are not likely
to be propagated to the hand-rolled items.