mktables: Fix Useless use of const in void context
The code was attempting to get a slice of an entire array excluding the
0th element.  But it was using the wrong incantation.

The array is the path to a file with each element being one directory.

And it just happened to work because of the current layout of the
directories.

This commit moves this slice fetch so it is not done in the loop, adds
another sanity check, and is generally clearer what is going on.