perl.h: Remove obsolete comment
This comment is not true; it may have been true once.  assert.h
is always included, and assert() is always supposed to be defined, since
at least C89.  But the code acts like it is possible to not be defined;
it's not a big deal to retain that check, so I'm leaving it in.