Perl_assert() expands to noop if Coverity, etc,
Some assertions have caused Coverity to complain in the past, or
overflowed for some platforms, so the definition of __ASSERT_ in handy.h
expands to a no-op if under Coverity or the system doesn't have a normal
sized macro buffer.  Bring that definition here to also be the same for
Perl_assert().