pp_pack: utf8_to_byte: Use 'uv' form, not 'uvchr'
Like the previous commit, this converts a call to utf8n_to_uvchr() to
instead use utf8_to_uv_flags().

And like that commit, the previous code thought it was checking for
failure and croaking when it occurred, but was checking the wrong way so
that it never could fail.  And now it does a proper check and can fail.