pp_pack: utf8_to_bytes: Use 'uv' form, not 'uvchr'
Like previous commits, this converts a call to utf8n_to_uvchr() to
instead use utf8_to_uv_flags().

Unlike those commits, the previous code did in fact properly check for
failure.  But upon failure, it semi-unsafely assumed the start byte was
accurate as to the intended byte length of the input character.  (It did
check that the returned value did not go past the end of the buffer, by
using UTF8_SAFE_SKIP() ).  The new function call always returns the
correct number of bytes, so just use that instead of trying to recompute
it.

The mechanism for deteriming failure no longer depends on the CHECK_ONLY
flag, so that is removed.