pp_pack.c: S_utf8_to_bytes: Rmv loop
Prior to this commit, the code went through the buffer with warnings
turned off, but saving the fact if there were problems.  If so, it ran
through the buffer again, with warnings turned on to display those
problems.

I'm unsure of why it didn't output any warnings until it got to the end.
But utf8_to_uv_msgs() allows us to store up the messages and output them
later in the same order as encountered, and without having to reparse