pp_pack: unpack: Use 'uv' form, not 'uvchr'
Like previous commits, this converts a call to utf8n_to_uvchr() to use
instead utf8_to_uv_flags().

And like previous commits, the previous code thought it was checking for
failure and croaking, but was checking the wrong way so that it never
could fail.  With the new commit, it can fail.

Also, the previous commit stored the result in an 'int', whose range can
be significantly less than the code point returned.  The value was
truncated without notice.  Now, I croak if the value got truncated.

This code doesn't fully make sense to me, as this is for unpacking a 'c'
format, which is supposed to be a signed 8-bit value, and what is
getting stored is a full integer value