pp_pack: unpack: Use 'uv' form, not 'uvchr'
This converts a call to utf8n_to_uvchr(), using instead utf8_to_uv().
The flag parameter UTF8_ALLOW_DEFAULT before this commit is the default
for utf8_to_uv(), so it can be omitted.

Like previous commits, the previous code thought it was checking for
failure and croaking, but was checking the wrong way so that it never
could fail.  In the new commit, I had to remove that checking, because
it caused test suite failures.