Add ASSUMEs for UTF-8 byte lengths
The maximum number of bytes in a Perl extended UTF-8 character is 13 on
ASCII platforms; 14 on EBCDIC.  Yet the variable that returns that
number is a Size_t in the cases changed by this commit.  By adding these
ASSUMES to these functions, the compiler may be able to do some
optimizations.

I looked through the code base, and found no other instances where such
a small value could be stored in a fully wide variable.

With link time optimization, an ASSUME may be helpful even in non-inline
functions.