Move prototype definition of SvPV_helper to embed.fnc
It's usually a bad idea to try to work around a limitation in common
code by copy-pasting and then modifiying to taste.  Fixes/improvements
to the common code rarely get propagated to the outlier.

I wrote code in 1ef9039bccb that did just this for the prototype
definition of SvPV_helper, because the place where it really belongs,
embed.fnc, couldn't (and still doesn't) handle function pointers as
arguments (patches welcome).

I should have at least added a comment to the common code noting the
existence of this outlier.

It turns out that that limitation can be worked around by declaring a
typedef of the pointer, and then using that in embed.fnc.

That's what this commit does.

This commit removes the final instance of duplicating the work of
embed.fnc in the core, except for some in the regex system whose
comments say the reason is to avoid making a typedef public.  I haven't
investigated these further.