Force SvPV_helper to be always inline
(when the compiler cooperates, that is)

I compiled toke.s using -O0 and looked at the output.  It had a full
copy of SvPV_helper in it.  But this function is called with various
constants that make much of the code dead when constant folded.  That
was not being done here.

This commit changes to force inlining.  Looking at a revised toke.s, I
saw that SvPV_helper was inserted multiple times in the code, and that
each occurrence had significantly fewer instructions than the original.
I infer that the compiler did the constant folding and pruned out the
dead code.