don't blame $x if reftype($x)/refaddr($x) is undef
When the "Use of uninitialized value" warning tries to find the source
of the problem, it skips over builtin::reftype/builtin::refaddr
operations and blames their arguments:

    $ perl -wE 'my $x = 42; $_ = reftype($x) eq ""'
    Use of uninitialized value $x in string eq at -e line 1.
    $ perl -wE 'my $x = 42; $_ = refaddr($x) eq ""'
    Use of uninitialized value $x in string eq at -e line 1.

This is wrong because $x is clearly defined.

This patch teaches S_find_uninit_var that reftype/refaddr can return
undef even if their arguments are perfectly defined.

Fixes #19273.