Add detail to -Dy debugging
Commit 6ceb4087860c6ef8e86e0c252feb738d635e9e3f added a way to cleanly
output UTF-8 tr/// values.  This commit uses that to improve the debug
output of compiling and running tr///.

For a simple tr of of transliterating Greek capital letters to
lowercase, the output of 'perl -Dy' has these added lines:

 > op.c: 6553: Compiling tr/*t/*r/; /c=0; /d=0; /s=0
 > *t is '\x{391}-\x{3a9}'
 > *r is '\x{3b1}-\x{3c9}'

Before the aforementioned commit the minus sign indicating a range would
not have rendered properly; so things like that were omitted from the
debug output.

The output also now includes special mention of the special casing where
the input is complemented, and/or some characters not being translated
or get deleted.