op.c: Change a -Dy line to yv; and another vice versa
While debugging this code, I found that the final generated map was
unneeded and too large for most instances; so change that to not be
output without the verbose option.

Conversely, I found that I needed to often see the state of things after
the first pass, so change to make that not verbose.