Add valid_utf8_to_uv()
This is identical to valid_utf8_to_uvchr(). They are both internal
functions designed for when you are certain that the utf8 string to be
translated is well formed; generally you created it yourself earlier.

The only reason for this new synonym is to lessen the cognitive load on
programmers who should be using the "_uv" suffix functions, and not the
"_uvchr" suffix ones for these sorts of tasks. By having this synonym,
one doesn't have to learn that there are two.