Add STATIC_ASSERT_EXPR
This new type of a STATIC_ASSERT is used where an expression can go;
hence potentially in macro expansions.  Since it is compile-time, it
can't be used on arguments that are expressions with side effects, so
can safely be used in macros.

It has limitations however, only reaching full parity when gcc brace
groups are enabled, turning into an ASSUME which is a no-op in the worst
case.

However, most code will get compiled on a platform with gcc brace
groups, and will pass or fail at compile time; so the chances of
something bad actually getting into the field are minimal.