XS: INTERFACE: support perl package names
XS currently allows a Perl package name to be used where a C type is
expected; behind the scenes, s/:/_/g is performed on the type string.

For example this XS declaration,

    X::Y
    foo(...)

causes this C declaration to be emitted for RETVAL:

    X__Y RETVAL;

However, in the presence of the INTERFACE keyword, e.g.

    X::Y
    foo(...)
      INTERFACE: ....

an extra variable to hold a function pointer is declared and
initialised, but the return type of the function gets declared as X::Y
rather than X__Y, resulting in uncompilable C code being generated.

This commit fixes that.

Before, the code emitted was:

    dXSFUNCTION(X::Y);
    XSFUNCTION = XSINTERFACE_FUNC(X::Y,cv,XSANY.any_dptr);

but is now:

    dXSFUNCTION(X__Y);
    XSFUNCTION = XSINTERFACE_FUNC(X__Y,cv,XSANY.any_dptr);