Consistently use `size_t` as array count/index in OP_MULTIPARAM
Recently-added code was written largely by copying existing practices in
various places. In particular, much of the original code was using `UV`
typed variables to store sizes and indexes in various list or array
structures, counts of parameters, and so on.

On fully 64-bit platforms this is all fine, but on 32-bit platforms with
-Duse64bitint enabled, this is a 64-bit integer that won't ever be
counting that high. These values might as well be stored as 32-bit
integers in that case.