pack: allocate msgs AV before pushing to it
This is an oversight from commit b0c3c5f582be. We need to make sure msgs
is not NULL before we can start pushing values.

Fixes a Coverity report:

    324         if (msgs) {
    >>>     CID 584148:         Possible Control flow issues  (DEADCODE)
    >>>     Execution cannot reach this statement: "while (Perl_av_count(my_per...".
    325             while (av_count(msgs) > 0) {

(Coverity thinks the code is unreachable because it is guarded by an 'if
(msgs)' statement, and 'msgs' is initialized to NULL and never modified
afterwards. In reality, it would probably crash in av_push() if any
warnings were to be generated.)