Remove v flag from embed.fnc
The flag was used to specify which functions are of a certain class, and
used to calculate a list to be sorted.  But that list is never going to
grow, so we can just specify it, already sorted.

The reason for doing this is mainly for a future commit which will want
the list in another program, and it's easier to not have to do the
calculations there.