deprecation.t: Comment out tests
We're superseding a warning about the impending fatalization of
goto-label-jump with an exception.  Hence the deprecation warning no
longer needs to be tested in t/porting/deprecation.t.

However, we may (or may not) have been using that particular warning as
corpur for *other* tests of deprecation warnings.  For the time being,
we'll comment out 3 test blocks and mark them as TODO.