Change names in OpTYPE_set() to avoid possible collisions
This uses the PERL_UNIQUE_NAME() macro created in the previous commit to
get unique names for variables in macro expansions.

Suppose you have

    #define foo STMT_START { ... } STMT_END

block, and have a variable 'bar' declared in it.  By first saying

    #define bar PERL_UNIQUE_NAME(bar)

'bar' will be replaced in the expansion of 'foo' by a unique value.

It is your responsibility to not have two 'bar' symbols in the same
program.  But if you forget, you will be warned by the compiler that
'bar' is redefined, so you can fix it
.