Fix two flaws in Configure
Configure conflated the file 'cppstdin' and the variable $cppstdin,
failing to initialize the latter.

It also overrode the hints file wrapper for feeding stdin to the C
preprocessor.  It did this without asking, but ran a simple test, and if
it passed, did the override.  This behavior dates to the mid 1990s.  The
problem is that on z/OS the simple test passes, but not all inputs do.

Some system header files on z/OS have C trigraphs, which cause modern
compilers to warn about them.  The hints file wrapper, which dates from
the late 1990's, did an "fgrep -v" to get rid of them, but it doesn't
end up getting used.  The simple test that Configure ends up of course
would not think to try trigraphs.  (There may be other issues as well,
but this one really pops out.)

This commit initializes $cppstdin, and doesn't try to override a
furnished cppstdin from the hints.