Make croak() always expand to Perl_croak_nocontext()
Perl almost always opts for saving time over saving space.  Hence, we
have croak() that saves time at the expense of space, but needs thread
context available; and croak_no_context() that doesn't need that, but
takes extra time

But, when we are about to die, time isn't that important.  Even if we
are doing eval after eval in a tight loop, the potential time savings of
passing the thread context to Perl_croak is insignificant compared to
the tear-down that follows.  My claim then is that croak() never needed
a thread context parameter to save a bit of time just before death.  It
is an optimization that isn't worth it.  And having it do so required
the invention of croak_nocontext(), and the extra cognitive load
associated with two methods for the same task.

This commit changes plain croak() to not use a thread context parameter.
It and croak_nocontext() now behave identically.  That means that going
forward, people will likely choose croak() which requires less typing
and occupies fewer columns on the screen, and they won't have to
remember which form to use when.