t/op/magic.t redirect stderr to /dev/null
This file calls ps(1), assuming that ps accepts the 'h' and 'hc'
options.  This is non-portable.  The comments in the code indicate that
these tests are optional; just skip them if the ps doesn't work.  The
problem is that noise gets generated on such platforms.  Simply redirect
stderr to /dev/null, suppressing any warnings.