ParseXS: add tests for standard_typemap_locations
Add some more tests to

    ExtUtils-ParseXS/t/101-standard_typemap_locations.

In particular, this time with a faked-up @INC. This is mainly
so that the next few commits will only change the things we expect
when updating standard_typemap_locations().

Note that this new test directly tests for expected strings such as
'../../typemap', rather than using File::Spec->updir() etc. This is to
avoid the test code being essentially a duplicate of the sub code. It
skips this test on platforms which don't seem to use '..' etc.